+++ changelog.md
@@ -9,12 +9,14 @@ Codion Change Log
 - FrameBuilder.componentListener() added.
 ### is.codion.framework.json.domain
 - EntityObjectMapper, methods for registering procedure and function argument and return types added, for serialization.
+- EntityObjectMapper, method for registering report parameter type added, for serialization.
 ### is.codion.framework.db.core
 - EntityConnection.report() reportParameters renamed parameter.
 ### is.codion.framework.servlet
 - EntityService.FunctionHandler and ProcedureHandler.json() added.
 - EntityService.deserialize() removed, Serializer.deserialize() used instead.
 - EntityService.JSON (default true) and SERIALIZATION (default false) configuration properties added, other properties renamed.
+- EntityService.ReportHandler.json() added.
 ### is.codion.framework.db.http
 - JsonHttpEntityConnection, JSON based function and procedure execution added.
 - HttpEntityConnection.JSON default back to being true.